Persistence to CSV files
* Persistence to CSV files

Replaces persistence of entities to a single json file with persistence to multiple csv files.
Changes the collections of entities to be read-only.

* Changes in catalog process execution

Removes cancellation token. Prevents the app from closing if the catalog process is running. Divides the catalog process in batches to allow the user to close the app even while cataloging large photo collections.